requirements.txt
Flask==2.3.3
Werkzeug==3.0.1
mysql-connector-python==8.0.33
